creating the match interface class.
the matchs show them self more complex what they look like. To create the matchs as should to be, we are rename some files and changing a lot the matchs and rules classes.

this commit create the match interface class what define, in a very generic mode, how any match should work.

git-svn-id: http://subversion.assembla.com/svn/codetodiagram@120 080cfd25-d5d9-4c59-82ea-70d02a200d8b